// src/types.ts
export type Lang =
  | 'en'
  | 'fr'
  | 'es'
  | 'pt'
  | 'pt-BR'
  | 'de'
  | 'nl'
  | 'cs'
  | 'hr'
  | 'sv'
  | 'no'
  | 'ar'
  | 'fa'
  | 'tr'
  | 'uz'
  | 'ja'
  | 'ko'
  | 'zh'
  | 'zh-TW'
  | 'it'
  | 'id'
  | 'ru'
  | 'uk'

export type LocalizedName = Partial<Record<Lang, string | null>>
// ...
export interface BaseCamp {
  city: string | null
  facility?: string | null
  country?: string | null
  lat?: number
  lon?: number
}

export interface Team {
  code: string
  fifaId: string | null
  group: string
  name: LocalizedName
  iso2: string | null
  ranking: number | null
  rankingPrev: number | null
  baseCamp: BaseCamp | null
  colors: string[]
  nickname: string | null
  web: string | null
  flag: string
}
// ...
export type PosBucket = 'GK' | 'DF' | 'MF' | 'FW'

export interface SquadPlayer {
  id: string
  no: number | null
  pos: PosBucket
  name: string
  dob: string | null
  caps: number | null
  goals: number | null
  club: string | null
  clubNat: string | null
  captain: boolean
  wiki: string | null // English Wikipedia article URL
}